Add "{clearOnly: true}" option to forceCreateSheet.
From the IntelliSense description:
clearOnly: If the sheet already exists, keep it as is, and only clear its grid. This results in a faster operation, and avoid a screen-update flash (and the re-setting of the current selection).
Note that clearing the grid does not remove floating objects like charts, so it's more of a "soft" operation, rather than a complete forceful re-creation of a sheet.

Also a tiny bit of refactoring, and removing/changing stale JSDOC comments.